btc/requirements.txt
ccxt==4.2.18
pandas==1.5.3
numpy==1.24.3
ta-lib==0.4.28
flask==2.3.3
flask-socketio==5.3.6
requests==2.31.0
python-dotenv==1.0.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
websocket-client==1.6.4
schedule==1.2.0